requirements.txt
flask==2.2.5
flask-cors==4.0.0
firebase-admin==6.2.0
requests==2.31.0
paho-mqtt==1.6.1
gunicorn==21.2.0
